requirements.txt
fastapi
uvicorn[standard]
pydantic
pymongo
python-dotenv
requests
deep-translator
google-generativeai
openai
